Move LineBrush into public Consolonia.Controls namespace
