+++ src/Consolonia.Core/Drawing/PixelBufferImplementation/DrawingBoxSymbol.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
+using Consolonia.Controls;
 
 namespace Consolonia.Core.Drawing.PixelBufferImplementation
 {
