+++ src/Consolonia.Gallery/Gallery/GalleryViews/GalleryBorders.axaml
@@ -1,6 +1,6 @@
 <UserControl xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:drawing="clr-namespace:Consolonia.Core.Drawing;assembly=Consolonia.Core"
+             xmlns:consolonia="clr-namespace:Consolonia.Controls;assembly=Consolonia.Core"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
@@ -15,7 +15,7 @@
             </Border>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="SingleLine"
+                    <consolonia:LineBrush LineStyle="SingleLine"
                                        Brush="Green" />
                 </Border.BorderBrush>
                 <TextBlock Text="LineStyle=SingleLine"
@@ -23,7 +23,7 @@
             </Border>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="DoubleLine"
+                    <consolonia:LineBrush LineStyle="DoubleLine"
                                        Brush="Blue" />
                 </Border.BorderBrush>
                 <TextBlock Text="LineStyle=DoubleLine"
@@ -31,7 +31,7 @@
             </Border>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="Edge"
+                    <consolonia:LineBrush LineStyle="Edge"
                                        Brush="DarkGreen" />
                 </Border.BorderBrush>
                 <TextBlock Text="LineStyle=Edge"
@@ -39,7 +39,7 @@
             </Border>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="EdgeWide"
+                    <consolonia:LineBrush LineStyle="EdgeWide"
                                        Brush="Purple" />
                 </Border.BorderBrush>
                 <TextBlock Text="LineStyle=EdgeWide"
@@ -47,7 +47,7 @@
             </Border>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="Bold"
+                    <consolonia:LineBrush LineStyle="Bold"
                                        Brush="DarkGray" />
                 </Border.BorderBrush>
                 <TextBlock Text="LineStyle=Bold"
@@ -57,7 +57,7 @@
         <WrapPanel>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="Edge"
+                    <consolonia:LineBrush LineStyle="Edge"
                                        Brush="Gray" />
                 </Border.BorderBrush>
                 <ComboBox Width="10">
@@ -70,7 +70,7 @@
             <Border BorderThickness="1 0 1 1"
                     Padding="0 1 0 0">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="Edge"
+                    <consolonia:LineBrush LineStyle="Edge"
                                        Brush="Gray" />
                 </Border.BorderBrush>
                 <TextBlock Text="Partial borders"
@@ -80,7 +80,7 @@
         <WrapPanel>
             <Border BorderThickness="1">
                 <Border.BorderBrush>
-                    <drawing:LineBrush LineStyle="Edge"
+                    <consolonia:LineBrush LineStyle="Edge"
                                        Brush="Gray" />
                 </Border.BorderBrush>
                 <Panel Background="AntiqueWhite">
