+++ src/Consolonia.Themes/Templates/Controls/DialogWindow.axaml
@@ -2,7 +2,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:controls="clr-namespace:Consolonia.Core.Controls;assembly=Consolonia.Core"
         xmlns:helpers="clr-namespace:Consolonia.Themes.Templates.Controls.Helpers"
-        xmlns:drawing="clr-namespace:Consolonia.Core.Drawing;assembly=Consolonia.Core">
+        xmlns:consolonia="clr-namespace:Consolonia.Controls;assembly=Consolonia.Core">
     <Styles.Resources>
         <Thickness x:Key="ConsoloniaDialogWindowBorderMargin">1,0</Thickness>
     </Styles.Resources>
@@ -46,7 +46,7 @@
                             <Border BorderThickness="{TemplateBinding BorderThickness}"
                                     Margin="{DynamicResource ConsoloniaDialogWindowBorderMargin}">
                                 <Border.BorderBrush>
-                                    <drawing:LineBrush
+                                    <consolonia:LineBrush
                                         Brush="{Binding Path=BorderBrush, RelativeSource={RelativeSource Mode=TemplatedParent}}"
                                         LineStyle="DoubleLine" />
                                 </Border.BorderBrush>
